Document service refactor
